CustomCNN implemented
