+++ catanatron_experimental/catanatron_experimental/machine_learning/players/ppo.py
@@ -1,5 +1,6 @@
 from typing import Iterable
 import numpy as np
+import os
 from sb3_contrib import MaskablePPO
 
 from catanatron.game import Game
@@ -7,6 +8,11 @@ from catanatron.models.actions import Action
 from catanatron.models.player import Player
 from catanatron_gym.envs.catanatron_env import from_action_space, to_action_space
 from catanatron_gym.features import create_sample, get_feature_ordering
+from catanatron_gym.board_tensor_features import (
+    create_board_tensor,
+    is_graph_feature,
+)
+from catanatron_experimental.machine_learning.custom_cnn import CustomCNN
 
 class PPOPlayer(Player):
     """
@@ -16,25 +22,22 @@ class PPOPlayer(Player):
     def __init__(self, color, model_path=None):
         super().__init__(color)
         self.model = None
+        self.numeric_features = None
         if model_path is None:
-            import os
             script_dir = os.path.dirname(os.path.abspath(__file__))
             model_path = os.path.join(script_dir, '..', 'model.zip')
         if model_path:
             self.load(model_path)
 
-    def __getstate__(self):
-        state = self.__dict__.copy()
-        state['model'] = None
-        return state
-
-    def __setstate__(self, state):
-        self.__dict__.update(state)
-        self.model = None
-
     def decide(self, game: Game, playable_actions: Iterable[Action]):
         if self.model is None:
             raise ValueError("Model not loaded. Call load() first.")
+
+        # Initialize numeric_features based on the current game
+        if self.numeric_features is None:
+            num_players = len(game.state.players)
+            self.features = get_feature_ordering(num_players)
+            self.numeric_features = [f for f in self.features if not is_graph_feature(f)]
 
         # Generate observation from the game state
         obs = self.generate_observation(game)
@@ -55,12 +58,20 @@ class PPOPlayer(Player):
             return list(playable_actions)[0]
 
     def generate_observation(self, game: Game):
-        # Create the observation using create_sample
+        # Create the sample
         sample = create_sample(game, self.color)
-        # Get the list of features used
-        features = get_feature_ordering(num_players=len(game.state.players))
-        # Create the observation vector
-        obs = np.array([float(sample[i]) for i in features], dtype=np.float32)
+
+        # Generate board tensor
+        board_tensor = create_board_tensor(game, self.color, channels_first=True).astype(np.float32)
+
+        # Extract numeric features
+        numeric = np.array([float(sample[i]) for i in self.numeric_features], dtype=np.float32)
+
+        # Create the observation
+        obs = {
+            'board': board_tensor,
+            'numeric': numeric
+        }
         return obs
 
     def generate_action_mask(self, playable_actions: Iterable[Action]):
@@ -87,4 +98,9 @@ class PPOPlayer(Player):
             raise ValueError(f"Action {action} not in playable actions.")
 
     def load(self, path):
-        self.model = MaskablePPO.load(path)
+        self.model = MaskablePPO.load(
+            path,
+            custom_objects={
+                'features_extractor_class': CustomCNN,
+            }
+        )
