+++ catanatron_experimental/catanatron_experimental/machine_learning/train_ppo_agent.py
@@ -6,13 +6,11 @@ import atexit
 
 import gymnasium as gym
 import torch as th
-from torch import nn
 import numpy as np
 from stable_baselines3.common.callbacks import (
     CheckpointCallback,
     CallbackList,
 )
-from stable_baselines3.common.torch_layers import BaseFeaturesExtractor
 from sb3_contrib.common.maskable.policies import MaskableActorCriticPolicy
 from sb3_contrib.common.wrappers import ActionMasker
 from sb3_contrib.ppo_mask import MaskablePPO
@@ -20,6 +18,7 @@ import wandb
 from wandb.integration.sb3 import WandbCallback
 
 from catanatron import Color
+from catanatron_experimental.machine_learning.custom_cnn import CustomCNN
 from catanatron_experimental.machine_learning.players.value import (
     ValueFunctionPlayer,
 )
@@ -65,51 +64,9 @@ def learning_rate_schedule(initial_lr, final_lr):
         return final_lr + (initial_lr - final_lr) * progress_remaining
     return lr_schedule
 
-class CustomCNN(BaseFeaturesExtractor):
-    """
-    Custom CNN to process the board observations.
-    :param observation_space: (gym.Space)
-    :param cnn_arch: List of integers specifying the number of filters in each Conv layer.
-    :param features_dim: (int) Number of features extracted.
-    """
-
-    def __init__(
-        self,
-        observation_space: gym.spaces.Dict,
-        cnn_arch,
-        features_dim: int = 256,
-    ):
-        super(CustomCNN, self).__init__(observation_space, features_dim)
-        n_input_channels = observation_space["board"].shape[0]
-
-        layers = []
-        in_channels = n_input_channels
-        for out_channels in cnn_arch:
-            layers.append(nn.Conv2d(in_channels, out_channels, kernel_size=3, stride=1, padding=1))
-            layers.append(nn.BatchNorm2d(out_channels))
-            layers.append(nn.ReLU())
-            in_channels = out_channels
-        layers.append(nn.Flatten())
-        self.cnn = nn.Sequential(*layers)
-
-        # Compute the number of features after CNN
-        with th.no_grad():
-            sample_board = th.as_tensor(observation_space.sample()["board"][None]).float()
-            n_flatten = self.cnn(sample_board).shape[1]
-
-        n_numeric_features = observation_space["numeric"].shape[0]
-        self.linear = nn.Sequential(
-            nn.Linear(n_flatten + n_numeric_features, features_dim), nn.ReLU()
-        )
-
-    def forward(self, observations: dict) -> th.Tensor:
-        board_features = self.cnn(observations["board"])
-        concatenated_tensor = th.cat([board_features, observations["numeric"]], dim=1)
-        return self.linear(concatenated_tensor)
-
 def main():
     # ===== Params:
-    total_timesteps = 10_000
+    total_timesteps = 10_000_000
     cnn_arch = [64, 64, 32]
     net_arch = [dict(vf=[512, 256, 128], pi=[512, 256, 128])]
     activation_fn = th.nn.Tanh
@@ -121,7 +78,7 @@ def main():
     map_type = "BASE"
     enemies = [ValueFunctionPlayer(Color.RED)]
     reward_function = build_partial_rewards(vps_to_win)
-    representation = "vector"
+    representation = "mixed"
     batch_size = 8192
     gamma = 0.98
     normalized = False
@@ -197,12 +154,13 @@ def main():
     device = th.device('cuda' if th.cuda.is_available() else 'cpu')
     print(f"Using device: {device}")
 
-    path = "catanatron_experimental/catanatron_experimental/machine_learning/model.zip"
+    script_dir = os.path.dirname(os.path.abspath(__file__))
+    model_path = os.path.join(script_dir, 'model.zip')
     try:
-        model = MaskablePPO.load(path, env, device=device)
+        model = MaskablePPO.load(model_path, env, device=device)
         print("Loaded", "model.zip")
     except Exception as e:
-        print(f"Failed to load the model from {path}: {e}")
+        print(f"Failed to load the model from {model_path}: {e}")
         print("Creating a new model.")
         policy_kwargs: Any = dict(activation_fn=activation_fn, net_arch=net_arch[0])
         if representation == "mixed":
@@ -240,12 +198,12 @@ def main():
                 total_timesteps=int(total_timesteps / selfplay_iterations),
                 callback=callback,
             )
-            model.save(path)
+            model.save(model_path)
     else:
         model.learn(total_timesteps=total_timesteps, callback=callback)
-        model.save(path)
+        model.save(model_path)
 
-    model.save(path)
+    model.save(model_path)
     run.finish()
 
 if __name__ == "__main__":
